refactored tooltip ui
